Use probe_dns_duration_seconds
This PR:

https://github.com/grafana/synthetic-monitoring-agent/pull/54

uses an updated version of blackbox_exporter that exposes a metric
called "probe_dns_duration_seconds", which is the time it takes for DNS
check to complete from connection to getting back the RR from the DNS
server. The metric has a label "phase" with three different values for
each phase of the query, just like probe_http_duration_seconds.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>