Validate PING, DNS and TCP targets
Add validations for the targets of PING, DNS and TCP checks. Make sure
we are getting something that is either a valid IP address or a valid
hostname. The API will also perform this validation, possibly more
strictly, so this is only meant to give the user some immediate
feedback.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>