fix: use new metric names
In https://github.com/grafana/synthetic-monitoring-agent/pull/73 there's
a change in the way that derived metrics are created. Specifically,
names like "probe_success_sum" change to "probe_all_success_sum" to
indicate that this metric contains information about _all_ the runs. The
_sum and _count metrics reflect the fact that they keep a running sum
and running count. We use those instead of simply probe_success in order
to avoid relying on the existence (or not) of the metric in order to
detect failures. The problem with that naming is that it will clash (in
the future) with the name of the original metric, probe_success in this
case. In order to avoid that, we rename the derived metric to
probe_all_success.

This change depends on that PR being merged _and_ deployed.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>