chore: reduce maximum Jest workers
Jest is too aggressive about running tests in parallel. It seems there's
something about the CheckEditor test suite that causes it to use more
memory than the other tests, so when running on all CPUs, it chews up
memory like nothing.

Reduce maxWorkers to 25% of the available CPUs as a guess... on my
machine, going above 25% does not reduce the test time (and in fact
going above 50% starts to increase it). I would happily keep it at 1 or
2, as that doesn't increase the total run time by any significant
amount.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>