fix: legend displays "Value" instead of probe name
In the DNS dashboard, the query is wrong, as it should be summing by
probe and not phase to coincide with the panel's description. In the
same panel, the legend should use the probe label, not the phase.

The the HTTP dashboard, the query is correct but the legend has the same
issue. Same goes for ping and TCP dashboards.

Also fix some casing issues in the panel titles.

Fixes #146.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>