Feature: add "reachability" to dashboards
Add a "reachability" panel to all dashboards (and a reachability column
to the summary). This reflects the fact that some probes might fail to
reach a given target because of network errors between the probe and the
target, not associated to either of them.

Also redefine uptime a measure of how much time the target was reachable
by any probe vs unreachable from all probes.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>