Feature: add filtering by custom labels
This adds dropdown menus to the summary dashboard that allow the user to
select a set of three labels to use as filters for the dashboard.

Each filter is composed of _two_ dropdowns: the first one specifies the
label to use for filtering and the second one the corresponding value.

There's a fourth, hidden, dropdown that is used to compute the effective
value of the "check type" dropdown, after applying the filters, so that
the dashboard doesn't show rows with empty panels.

The defaults in the dashboard are such that all the checks are selected.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>