Default server for DNS to domain name
Currently the default server for DNS checks is set to an IPv4 address.
But users can choose to use IPv6 instead, at which point the IPv4
default address leads to the check constantly failing. This default will
work for both IPv4 and IPv6 without issue.

See this working in practice on this check here:
https://syntheticmonitoring.grafana.net/a/grafana-synthetic-monitoring-app/?page=checks&id=63Revert "Revert "Default server for DNS to domain name""