fix: bump go version down
This makes it easier for people to get started. Once Go 1.22 is more
widespread, it should become easier, as it will download the required
version automatically (unless of course you are me and you go like "oh
no, not today and not ever!", in which case you know how to fix this).

Bingo created gum.mod with 'go 1.22.5' because that's what I have right
now, but gum itself only wants 1.21.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>