fix: rework logic to locate bingo
Try to make less assumptions as to what's available and what isn't. If
there's a bingo command on the path, assume it's what we want. If not,
try to figure out if GOBIN is in the user's PATH, and if not, reach to
that location directly, with a reasonable fallback.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>