fix: make sure gum is added
`bingo get` will get all the tools assuming `.bingo` is part of the
repository already. If that's not the case, it will initialize bingo,
but it won't actually do anything and therefore .bingo/variables.env
doesn't get created.

Assume we don't have .bingo set up already, so do that by calling `bingo
get`, and then assume we don't have gum, so get it if the corresponding
mod file is missing.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>