Add release type
This type only exists to allow forcing release-please PR generation when
the commits from the previous release are all internal (none of them
comply with the supported sections by the release-please configuration).
The convention is to push a single empty commit with message:
`release: Internal release`
Which will indicate that that release only contains internal changes.