chore: Setup release-please (#933)
* chore: Setup release-please

* Simplify release-please configuration

Sets a minimal configuration for release-please to only handle feat,
fix, and docs commits. Other categories might be included in the future
if considered necessary.

* Update release-please manifest

* Add release type

This type only exists to allow forcing release-please PR generation when
the commits from the previous release are all internal (none of them
comply with the supported sections by the release-please configuration).
The convention is to push a single empty commit with message:
`release: Internal release`
Which will indicate that that release only contains internal changes.

* Support chore and refactor conventional commit types

These types are already being used in the workflow for the app and
included in the changelog and release notes, therefore it makes sense to
keep supporting them and not break the current pattern.