chore: Fix release CI/CD (#958)
After the change in the release process, the tag is now generated when
the release PR is merged to main. At this point, there is a current
trigger for a tag that tries to publish to gcom the released plugin.
However, this step should not happen automatically on the tag generation
but happen when promoting the release to production. Therefore, move
this step as the last step for the promote->production event trigger.