fix: Publish script
A space before the bash bang was making it being executed through sh in
ubuntu:latest as part of the promote production CI pipeline, on which an
option set in the script is not supported. Remove the extra space so the
sript is executed in bash and 'set -eufo pipefile' is applied correctly.